Add files via upload
Still hasn't figured out how to use bootstrap nor how to properly use <table> tags. But, I experienced the first taste of CSS power of changing colors and fonts.  Albeit, not advanced, it is a sweet start. :) 
Will have to learn more about how to design the layout of the pages.